+++ tests/data/app/controllers.php
@@ -79,6 +79,18 @@ class redirect_header_interval {
     }
 }
 
+class redirect_base_uri_has_path {
+    function GET() {
+        header('Refresh:0;url=/somepath/info');
+    }
+}
+
+class redirect_base_uri_has_path_302 {
+    function GET() {
+        header('Location: /somepath/info', true, 302);
+    }
+}
+
 class external_url {
     function GET() {
         include __DIR__ . '/view/external_url.php';
@@ -126,6 +138,13 @@ class cookiesHeader {
         header("Set-Cookie: a=b;Path=/;");
         header("Set-Cookie: c=d;Path=/;", false);
         include __DIR__.'/view/index.php';
+    }
+}
+
+class iframe {
+    public function GET()
+    {
+        include __DIR__.'/view/iframe.php';
     }
 }
 
@@ -193,7 +212,7 @@ class register {
     function GET() {
         include __DIR__.'/view/register.php';
     }
-    
+
     function POST() {
         $this->GET();
     }