+++ tests/data/app/index.php
@@ -31,6 +31,9 @@ $urls = array(
     '/relative/redirect' => 'redirect_relative',
     '/redirect_twice' => 'redirect_twice',
     '/relative/info' => 'info',
+    '/somepath/redirect_base_uri_has_path' => 'redirect_base_uri_has_path',
+    '/somepath/redirect_base_uri_has_path_302' => 'redirect_base_uri_has_path_302',
+    '/somepath/info' => 'info',
     '/facebook\??.*' => 'facebookController',
     '/form/(.*?)(#|\?.*?)?' => 'form',
     '/articles\??.*' => 'articles',
@@ -40,6 +43,7 @@ $urls = array(
     '/content-cp1251' => 'contentType2',
     '/unset-cookie' => 'unsetCookie',
     '/external_url' => 'external_url',
+    '/iframe' => 'iframe',
 );
 
 glue::stick($urls);