+++ tests/unit/Codeception/Module/TestsForWeb.php
@@ -1292,4 +1292,19 @@ abstract class TestsForWeb extends \PHPUnit_Framework_TestCase
         $form = data::get('form');
         $this->assertEquals('this & that', $form['name']);
     }
+
+    public function testSeeInDeactivatedField()
+    {
+        $this->module->amOnPage('/form/complex');
+        $this->module->seeInField('#disabled_field', 'disabled_field');
+        $this->module->seeInField('#salutation', 'mr');
+    }
+
+    public function testSwitchToIframe()
+    {
+        $this->module->amOnPage('/iframe');
+        $this->module->switchToIframe('content');
+        $this->module->see('Is that interesting?');
+        $this->module->click('Ссылочка');
+    }
 }