+++ CHANGELOG.md
@@ -2,9 +2,10 @@
 
 #### 2.1.3
 
+* [REST] Added matching data types by with new methods `seeResponseMatchesJsonType` and `dontSeeResponseMatchesJsonType. See #2391
+* Environment files can now be located in subfolders of `tests/_env` by @Zifius
 * [Symfony2] Fixed issue when accessing profiler when no request has been performed #652.
 * [Symfony2] Added amOnRoute and seeCurrentRouteIs methods to Symfony2 module, by @raistlin
-* [ZF1] Added amOnRoute and seeCurrentRouteIs methods to ZF1 module, by @Naktibalda
 * [ZF2] Added amOnRoute and seeCurrentRouteIs methods to ZF2 module, by @Naktibalda
 * Fixed issue with trailing slashes in `seeCurrentUrlEquals` and `dontSeeCurrentUrlEquals` methods #2324. By @janhenkgerritsen
 * Warning is displayed once using unconfigured environment.
@@ -20,8 +21,9 @@
 * [Soap] Fixed SOAP module #2296. By @relaxart
 * Fixed a bug where blank lines in a groups file would run every test in the project #2297. By @imjoehaines
 * [WebDriver] seeNumberOfElements should only count visible elements #2303. By @sascha-egerer
-* [Functional] Verbose output for redirects. By @Naktibalda
-* [Functional] Throw Codeception\Exception\ExternalUrlException when framework module tries to open an external URL #2328. By @Naktibalda
+* [PhpBrowser][Frameworks] Verbose output for all HTTP requests. By @Naktibalda
+* [PhpBrowser][Frameworks] Throw Codeception\Exception\ExternalUrlException when framework module tries to open an external URL #2328. By @Naktibalda
+* [PhpBrowser][Frameworks] Added switchToIframe method. By @Naktibalda
 
 #### 2.1.2